feat: userService register 메서드 작성

+++ core/core-domain/src/main/java/com/josh/joinus/core/dto/request/UserCreateRequest.java
@@ -10,5 +10,5 @@ public class UserCreateRequest {
     private String nickname;
     private Position position;
     private int careerYear;
-    private List<Integer> techIdList;
+    private List<Long> techIdList;
 }
